+++ app/src/main/java/com/example/myapplication/TempPlayer.java
@@ -17,12 +17,12 @@ public class TempPlayer extends GameObject {
     }
 
     @Override
-    public void update(int ms) {
+    public void update(long ms) {
 
     }
 
     @Override
-    public void render(Renderer renderer, int ms) {
+    public void render(Renderer renderer, long ms) {
         double[] position = getPosition();
         renderer.drawFilledCircle(position[0], position[1], Color.YELLOW, 10f);
     }
